requirements.txt
django
djangorestframework
psycopg2-binary
drf-yasg
python-decouple
Pillow
django-filter
python-slugify
djangorestframework-simplejwt
